Move message conversion to `stateful`
