+++ packages/engine/lib/stateful/src/message/mod.rs
@@ -2,6 +2,9 @@
 
 pub mod payload;
 
+//temporarily public
+pub mod arrow;
+
 mod inbound;
 mod kind;
 mod outbound;
